+++ routing/ann_validation.go
@@ -1,4 +1,4 @@
-package discovery
+package routing
 
 import (
 	"bytes"
